+++ axum/src/extract/extension.rs
@@ -10,7 +10,6 @@ use std::ops::Deref;
 ///
 /// ```rust,no_run
 /// use axum::{
-///     AddExtensionLayer,
 ///     extract::Extension,
 ///     routing::get,
 ///     Router,
@@ -31,7 +30,7 @@ use std::ops::Deref;
 /// let app = Router::new().route("/", get(handler))
 ///     // Add middleware that inserts the state into all incoming request's
 ///     // extensions.
-///     .layer(AddExtensionLayer::new(state));
+///     .layer(Extension(state));
 /// # async {
 /// # axum::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
 /// # };
@@ -57,7 +56,7 @@ where
             .get::<T>()
             .ok_or_else(|| {
                 MissingExtension::from_err(format!(
-                    "Extension of type `{}` was not found. Perhaps you forgot to add it?",
+                    "Extension of type `{}` was not found. Perhaps you forgot to add it? See `axum::extract::Extension`.",
                     std::any::type_name::<T>()
                 ))
             })
@@ -72,5 +71,19 @@ impl<T> Deref for Extension<T> {
 
     fn deref(&self) -> &Self::Target {
         &self.0
+    }
+}
+
+impl<S, T> tower_layer::Layer<S> for Extension<T>
+where
+    T: Clone + Send + Sync + 'static,
+{
+    type Service = crate::AddExtension<S, T>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        crate::AddExtension {
+            inner,
+            value: self.0.clone(),
+        }
     }
 }